requirements.txt
rxv
pySerial
simpleaudio
